Eliminating unnamed "main" function.
Besides the cloning of the travis-cache repository which is still done
when this script is sourced, this script now merely defines functions
that can be called from .travis.yml files. This change was made to make
it easier to define what this script actually does, and to promote
extensibility for future features. The README has also been updated
to reflect these changes.