Track removal of .subtree-clones in prepare-deploy
The prepare-deploy script was recently modified to no longer clone the
site-common and site-$site repositories into .subtree-clones. So the
nifty-script can no longer assume that it can work in those
repositories after calling prepare-deploy.

So this code is removed here.

The purpose of the removed code was to update the master branch within
site-common and site-$site. That task should now be performed directly
within prepare-deploy, (when deploying to staging because of a push to
the master branch of the sites repository).