Add a comment to document a non-obvious call to eval
As the comment describes, there is a bit of magic in the line that says:

	eval ${prepare_args/ /;}

At the time of this line of code, the value in $prepare_args will be
something like:

	DEPLOY_TAG=mode-user-XYZ DEPLOY_COMMON_TAG=site-mode-user-XYZ

And that syntax is suitable for putting directly onto a command line
so that the command will have those variables in its environment.

The purpose of the eval is to bring these variable assignments into
the current environment. To do that, we need two different statements,
and statements in the shell are separated by semicolons, not
spaces. So the bash syntax of ${var/ /;} replaces the space with a
semicolon and then the eval pulls in the variable assignments for the
current process's environment.