Revert "Expect a TRAVIS_TAG of "master" not "staging-*""
This reverts commit 5b252656df49f82337bcbc3fdc7bef8478cb0831.

The problem is that when Travis sees a push to site-<foo> on a branch
named "master" it doesn't know which commit of site-common to test
against. (Compare with a push of a tag named feature-travis-XYZ where
the site-common tag to checkout is foo-feature-travis-XZY).

This detail was missed originally because the checkout of the desired
tag for site-common is currently taking place in each of the
sites/foo/.travis.yaml files rather than in nifty-script directly.

For now, just back the buggy code out while we reconsider our options.