Redirect "make reqs" to /dev/null and call pip freeze
The output of "make reqs" was really bloating our travis logs. Not
only that, but id didn't even have useful information in it---it was
reporting only version numbers even when we installed a required
package from a git repository.

In contrast, the "pip freeze" output includes the git commit IDs for
any package installed via git.

So we discard the "make reqs" output and include the "pip freeze"
output instead.