nifty: Perform git-clone of site-common repository when testing site-foo
Previously, this code was duplicated in the .travis.yml file for each
of our sites. We can maintain one copy instead of N copies by hosting
it here in nifty-script.

In addition, this version now makes the checkout of a specific tag
conditional on the TRAVIS_TAG variable being set. This will allow
testing of commits to master, (in which case TRAVIS_TAG will be
unset).