Don't re-run tests if a .coverage directory already exists
With this change, nifty-using apps can arrange to have coverage
results generated during the testing process so that the nifty
verify_coverage_improvement needn't run the test suite again.

To do this, an app simply needs to:

    1. Defer the call to verify_coverage_improvement until after testing

    2. Change a call that looks something like:

		./manage.py test --keepdb

       to instead look something like:

		coverage erase
                coverage run $(COVERAGE_ARGS) ./manage.py test --keepdb

NOTE: Step (1) above is actually not possible for most of our apps
right now since they are actually calling the test_and_tag function
from .travis.yml and doing coverage verification after that could
semantically invalidate the tag already pushed, (in that it shouldn't
have been pushed if the coverage amount dropped).

See the next commit for an alternative.