Add a new nifty_test_app() function
This is very similar to test_and_tag() but with the following changes:

	It calls "make travis-tests" instead of "make test"
        It calls verify_coverage_improvement (which test_and_tag did not)

With these two changes, any app that has already migrated to have a
"travis-tests" target, and that also generates .coverage data during
the testing that that performs, can switch its "script" section of
.travis.yml to simply:

	- source .nifty/nifty-script
	- nifty_test_app

This will perform all testing and coverage verification without any
redundant running of the app's test suite.

For apps in private repositories, (those with a "source_key" defined
in .travis.yml), this function will also push out a "travis-tested"
tag if all tests succeed.