Add function to test separated platform repository.
With the inclusion of nimbis-platform into the sites repository, our nifty
script needs to be extended to also test the separated platform repository. This
function should be called by the travis script in the nimbis-platform
repository. Most of this function follows what would happen for a push to a
separated sites repository except without deployment, coverage testing, or
tagging. Piprot, make test, and caching are performed.