Test platform builds in house.
We add a check for if this build is for platform. If it is then perform testing
and return as there is no need for subtree pushing or deployment. This process
will help us deprecate the nimbis-platform repo and its ansible submodules
entirely. The idea is that we can work in sites and prevent the overhead of
working with submodules.