Rename "url" to "item" in __store_travis_cache
When we cache travis test runs we store a URL in the cache, (linking
to the previous Travis build results). But when we cache coverage
results we store a numerical value. Prior to this commit, the commit
message always said "URL" which was confusing.

In this commit, we switch to using "item" which works for both cases.