Fix nifty to be able to correctly use "set -e"
The trick for this is that the caller of nifty must actually invoke
nifty-script as a script, (and not source it). The problem with
sourcing was that things like "set -e" (or similar games with trap)
were leaking and interfering with Travis functionality that was
happening after the script invocation.

By invoking nifty-script as a script, nifty is running in a separate
sub-shell compared to the Travis script and the "set -e" behavior is
isolated to that.