Disable deploys to production
We want to be able to make releases (literally 'make release') with out
having this awake the Travis beast and deploying all over our production
server(s).

The releases are needed for our new deployment process, but we have yet
to take down the old servers.  Staging deploys will still work as
expected.