Start an X server (Xvfb) in the __init_nifty_script function
Previously, the Xvfb server was started when doing testing of sites,
but not when doing testing of apps. However, some apps, (such as
shop-core), are now wanting to use Selenium-based testing of an actual
web browser against a live Django instance, so we want to have the X
server available in all cases.