Append job number to the key used to cache test results
Without this job number, consider the following scenaroi:

    * A single build has multiple jobs, (via "matrix" in .travis.yml)

    * One of the jobs passes (and adds its URL to the cache)

    * One or more jobs fail, reporting that to the developer

    * The developer restarts the failed job

    * In the restart, nifty says "This code has been tested previously successfully"
      and points to the result of a different job.

We just encountered this problem in shop-core when using "matrix" to
test with different values of the TEST_BROWSER environment variable.

Note that we did not see this problem with "sites" because each job
within the build uses a different value for the "SITE" environment
variable and the cache key already includes that value.